page 형식 으로 리턴하는 데이터들을 PaginatedObject를 사용 하도록 리팩토링
PostCrudServiceReactiveMongoServiceImpl 작성

+++ src/main/kotlin/org/jeong/reservationinformation/common/util/PageUtil.kt
@@ -15,6 +15,6 @@ class PageUtil {
         return "/$startIndex/$endIndex/"
     }
 
-    fun getLastPage(totalCount: Long, size: Int): Long =
-            ceil(totalCount / size.toDouble()).toLong()
+    fun getLastPage(totalCount: Long, size: Int): Int =
+            ceil(totalCount / size.toDouble()).toInt()
 }
